move andrii controllers to separate folder and adding to them namespaces, installed vue.js
