+++ symfony/andrii/cauldron/templates/question/homepage.html.twig
@@ -1,4 +1,4 @@
-{% extends '@andrii/base.html.twig' %}
+{% extends '@andrii_cauldron/base.html.twig' %}
 
 {% block body %}
     <div class="jumbotron-img jumbotron jumbotron-fluid">
