+++ symfony/andrii/cauldron/templates/question/show.html.twig
@@ -1,4 +1,4 @@
-{% extends '@andrii/base.html.twig' %}
+{% extends '@andrii_cauldron/base.html.twig' %}
 
 {% block title %}Question: {{ question }}{% endblock %}
