+++ symfony/composer.json
@@ -62,7 +62,9 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "src/"
+            "App\\": "src/",
+            "Cauldron\\": "andrii/cauldron/src/",
+            "MVP_Office\\": "andrii/mvp_office/src/"
         }
     },
     "autoload-dev": {
