+++ symfony/config/packages/twig.yaml
@@ -1,4 +1,5 @@
 twig:
     default_path: '%kernel.project_dir%/templates'
     paths:
-        'andrii/templates': 'andrii'
+        '%kernel.project_dir%/andrii/cauldron/templates': andrii_cauldron
+        '%kernel.project_dir%/andrii/mvp_office/templates': andrii_mvp_office
