+++ symfony/config/services.yaml
@@ -20,12 +20,22 @@ services:
             - '../src/Entity/'
             - '../src/Kernel.php'
             - '../src/Tests/'
+    Cauldron\:
+        resource: '../andrii/cauldron/src/'
+    MVP_Office\:
+        resource: '../andrii/mvp_office/src/'
 
     # controllers are imported separately to make sure services can be injected
     # as action arguments even if you don't extend any base controller class
     App\Controller\:
         resource: '../src/Controller/'
         tags: ['controller.service_arguments']
+    Cauldron\Controller\:
+        resource: '../andrii/cauldron/src/Controller/'
+        tags: ['controller.service_arguments']
+    MVPOffice\Controller\:
+        resource: '../andrii/mvp_office/src/Controller/'
+        tags: [ 'controller.service_arguments' ]
 
     # add more service definitions when explicit configuration is needed
     # please note that last definitions always *replace* previous ones
