+++ symfony/package.json
@@ -6,7 +6,12 @@
         "core-js": "^3.0.0",
         "jquery": "^3.5.1",
         "regenerator-runtime": "^0.13.2",
+        "sass": "^1.32.6",
+        "sass-loader": "^10.0.0",
         "stimulus": "^2.0.0",
+        "vue": "^2.5",
+        "vue-loader": "^15",
+        "vue-template-compiler": "^2.6.12",
         "webpack-notifier": "^1.6.0"
     },
     "license": "UNLICENSED",
